minseok_unittest
lock, unlock, brake unittest
